Fixed T3D

+++ T3D/src/Actor.py
@@ -4,7 +4,7 @@ import torch.nn.functional as F
 
 
 class Actor(nn.Module): # TODO: Refactor this with Fatih's implementation
-    def __init__(self, state_size:int , action_size: int, max_action: int, hidden_size:int):
+    def __init__(self, state_size:int , action_size: int, max_action: float, hidden_size:int):
         super(Actor, self).__init__()
 
         self.fc1 = nn.Linear(state_size, hidden_size)
